lucas-shoes-master/requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.13.1
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.1.7
django-multiselectfield==0.1.9
django-taggit==1.1.0
django-weasyprint==0.5.4
django-widget-tweaks==1.4.5
html5lib==1.0.1
Pillow==6.2.1
pycparser==2.19
Pyphen==0.9.5
pytz==2019.3
six==1.12.0
sorl-thumbnail==12.5.0
tinycss2==1.0.2
WeasyPrint==50
webencodings==0.5.1
xlwt==1.3.0
